test(partial-encryption): split SOPS regression tests into focused modules
CodeScene flagged two test files over code-health thresholds after the #416 SOPS
regression tests landed:

- tests/integration/test_partial_encryption_e2e.py: 640 LoC (limit 600). Move the
  cohesive real-sops e2e class (TestSopsEncryptedSecretIsFullyEncryptedRealBinary)
  into a focused sibling module test_partial_encryption_sops_e2e.py, gated on the
  sops binary instead of dotenvx.
- tests/unit/test_partial_encryption.py: 79 functions (limit 75). Move the #416
  value-scan / fully-SOPS-encrypted-.secret family (has_plaintext_secret_value and
  the sops metadata-trailer tests, plus their _sops_ciphertext helper and
  _SOPS_ENCRYPTED_DOTENV fixture) into test_partial_encryption_sops.py.

Pure relocation: no assertions changed. All 74 unit + 19 integration partial-
encryption tests still pass. Each module is now within threshold (e2e well under
600 LoC; original unit module at 69 functions).
